# openapi.yaml
openapi: 3.0.3
info:
  title: Nutri Thata API
  description: API para plataforma de educação nutricional com sistema de módulos, convites e upload de mídia
  version: 1.0.0
  contact:
    name: API Support
    email: support@nutrithata.com
servers:
  - url: http://localhost:8080
    description: Desenvolvimento local
  - url: https://api.nutrithata.com
    description: Produção
security:
  - BearerAuth: []
paths:
...
  /auth/register/patient:
    post:
      tags:
        - Autenticação
      summary: Registrar novo paciente
      description: Registro público disponível apenas para role PATIENT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                  example: "João Silva"
                email:
                  type: string
                  format: email
                  example: "joao@email.com"
                password:
                  type: string
                  minLength: 8
                  example: "MinhaSenh@123"
                phone:
                  type: string
                  example: "(11) 99999-9999"
                birthDate:
                  type: string
                  format: date
                  example: "1990-01-01"
      responses:
        '201':
          description: Paciente registrado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          description: Email já está em uso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Dados de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags:
        - Autenticação
      summary: Fazer login
      description: Autenticação de usuário
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "joao@email.com"
                password:
                  type: string
                  example: "MinhaSenh@123"
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Dados de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags:
        - Autenticação
      summary: Renovar token JWT
      description: Gera novo token de acesso
      responses:
        '200':
          description: Token renovado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Token renovado"
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                        example: "eyJhbGciOiJIUzI1NiJ9..."
        '401':
          description: Token inválido ou expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      tags:
        - Autenticação
      summary: Obter usuário atual
      description: Retorna dados do usuário autenticado
      responses:
        '200':
          description: Dados do usuário
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        '401':
          description: Token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ========== USUÁRIOS ==========
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuthResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Login realizado com sucesso"
        data:
          type: object
          properties:
            token:
              type: string
              example: "eyJhbGciOiJIUzI1NiJ9..."
            user:
              $ref: '#/components/schemas/User'
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: "error"
        faultstring:
          type: string
          example: "Credenciais inválidas"
        stacktrace:
          type: string
          description: "Disponível apenas em ambiente de desenvolvimento"
          example: "com.nutrithata.exception.AuthenticationException: Invalid credentials..."

    # ========== MODELOS DE USUÁRIO ==========
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "João Silva"
        email:
          type: string
          format: email
          example: "joao@email.com"
        role:
          type: string
          enum: ["PATIENT", "PROFESSIONAL", "ADMIN"]
          example: "PATIENT"
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
...